Support account attributes

+++ lib/easy_auth/models/identities/oauth/base.rb
@@ -4,33 +4,46 @@ module EasyAuth::Models::Identities::Oauth::Base
   def self.included(base)
     base.class_eval do
       serialize :token, Hash
+      validates :username, :presence => true
       extend ClassMethods
     end
   end
 
   module ClassMethods
     def authenticate(controller)
-      oauth_token         = controller.params[:oauth_token]
-      oauth_verifier      = controller.params[:oauth_verifier]
-      access_token_secret = controller.session.delete('access_token_secret')
-      request_token       = OAuth::RequestToken.new(client, oauth_token, access_token_secret)
-      token               = request_token.get_access_token(:oauth_verifier => oauth_verifier)
-      username            = retrieve_username(token)
-      identity            = self.find_or_initialize_by_username username.to_s
-      identity.token      = {:token => token.token, :secret => token.secret}
-      account             = controller.current_account
+      if controller.params[:oauth_token].present? && controller.params[:oauth_verifier].present?
+        oauth_token         = controller.params[:oauth_token]
+        oauth_verifier      = controller.params[:oauth_verifier]
+        access_token_secret = controller.session.delete('access_token_secret')
+        request_token       = OAuth::RequestToken.new(client, oauth_token, access_token_secret)
+        access_token        = request_token.get_access_token(:oauth_verifier => oauth_verifier)
+        username            = retrieve_username(access_token)
+        identity            = self.find_or_initialize_by_username username.to_s
+        identity.token      = {:token => access_token.token, :secret => access_token.secret}
 
-      if identity.new_record?
-        account = EasyAuth.account_model.create(username_attribute => identity.username) if account.nil?
-        identity.account = account
+        unless controller.current_account && identity.account
+          account = EasyAuth.account_model.create!(account_attributes(access_token.params)) if account.nil?
+          identity.account = controller.current_account
+        end
+
+        identity.save!
+        identity
       end
-
-      identity.save!
-      identity
     end
 
-    def username_attribute
-      :email
+    def account_attributes(user_info)
+      setters = EasyAuth.account_model.instance_methods.grep(/=$/) - [:id=]
+      account_attributes_map.inject({}) do |hash, kv|
+        if setters.include?("#{kv[0]}=".to_sym)
+          hash[kv[0]] = user_info[kv[1]]
+        end
+
+        hash
+      end
+    end
+
+    def account_attributes_map
+      { :email => 'email' }
     end
 
     def new_session(controller)
@@ -49,10 +62,6 @@ module EasyAuth::Models::Identities::Oauth::Base
 
     def client_options
       { :site => site_url, :authorize_path => authorize_path }
-    end
-
-    def provider
-      raise NotImplementedError
     end
 
     def retrieve_username(token)
@@ -75,10 +84,6 @@ module EasyAuth::Models::Identities::Oauth::Base
 
     def site_url
       raise NotImplementedError
-    end
-
-    def scope
-      settings.scope
     end
 
     def client_id
